LayoutTests/platform/ios-simulator/transitions/transform-op-list-match-expected.txt

CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box'
Box should spin 540 degrees because that is the specified rotation and the transform operation lists match, so we should animate each operation separately.

PASS - "-webkit-transform.0" property for "box" element at 0.5s saw something close to: -0.70711
FAIL - "-webkit-transform.0" property for "box" element at 1s expected: 0 but saw: matrix(-1, 0.00000000000000036739403974420594, -0.00000000000000036739403974420594, -1, 0, 0)
